+++ frontend/src/features/consultants/components/profile/location-card.tsx
@@ -1,3 +1,6 @@
+import { SectionCard } from "../../../../components/shared/section-card";
+import { DetailField } from "../../../../components/shared/detail-field";
+
 interface LocationCardProps {
   readonly address1: string;
   readonly address2: string;
@@ -7,28 +10,7 @@ interface LocationCardProps {
   readonly postalCode: string;
 }
 
-function ProfileField({ label, value }: { 
-  readonly label: string; 
-  readonly value: string }) {
-  return (
-    <div className="flex flex-col" style={{ gap: "10px", minWidth: 0 }}>
-      <p
-        className="font-semibold"
-        style={{ color: "var(--color-text-secondary)", fontSize: "var(--text-h4)" }}
-      >
-        {label}
-      </p>
-      <p
-        className="font-semibold break-words"
-        style={{ color: "var(--color-text-primary)", fontSize: "var(--text-h3)" }}
-      >
-        {value || "—"}
-      </p>
-    </div>
-  );
-}
-
-function LocationCard({
+export default function LocationCard({
   address1,
   address2,
   suburb,
@@ -37,42 +19,20 @@ function LocationCard({
   postalCode,
 }: LocationCardProps) {
   return (
-    <div
-      className="bg-white rounded-2xl w-full flex flex-col"
-      style={{
-        padding: "28px 28px 28px 28px",
-        boxShadow: "0 2px 8px rgba(0,0,0,0.06)",
-        gap: "28px",
-      }}
-    >
-     
-      <h2
-        className="font-bold"
-        style={{ color: "var(--color-primary)", fontSize: "22px" }}
-      >
-        Location
-      </h2>
-
-      <hr style={{ borderColor: "var(--color-border)" }} />
-
+    <SectionCard title="Location">
       <div className="flex flex-col" style={{ gap: "28px" }}>
-        <ProfileField label="Address Line 1" value={address1} />
-        <ProfileField label="Address Line 2" value={address2} />
+        <DetailField label="Address Line 1" value={address1} variant="compact" />
+        <DetailField label="Address Line 2" value={address2} variant="compact" />
       </div>
 
       <hr style={{ borderColor: "var(--color-border)" }} />
 
-      <div
-        className="grid grid-cols-1 md:grid-cols-2"
-        style={{ gap: "28px" }}
-      >
-        <ProfileField label="Suburb"      value={suburb} />
-        <ProfileField label="City"        value={city} />
-        <ProfileField label="Province"    value={province} />
-        <ProfileField label="Postal Code" value={postalCode} />
+      <div className="grid grid-cols-1 md:grid-cols-2" style={{ gap: "28px" }}>
+        <DetailField label="Suburb" value={suburb} variant="compact" />
+        <DetailField label="City" value={city} variant="compact" />
+        <DetailField label="Province" value={province} variant="compact" />
+        <DetailField label="Postal Code" value={postalCode} variant="compact" />
       </div>
-    </div>
+    </SectionCard>
   );
 }
-
-export default LocationCard;
